pyproject.toml
[tool.poetry]
name = "trello-reports"
version = "0.2.0"
description = "CLI tool to generate reports from trello kanban boards"
authors = ["Rodrigo Machuca <rmachuca89@hotmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/rmachuca89/trello-reports"
keywords = ["trello"]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
py-trello-api = "0.20.0"
py-trello = "^0.20.1"

[tool.poetry.group.dev.dependencies]
nox-poetry = "^1.0.3"
flake8 = "^7.1.1"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
mypy = "^1.13.0"
black = "^24.10.0"
pre-commit = "^4.0.1"
nox = "^2024.10.9"
codecov = "^2.1.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
trello-reports = "trello_reports.cli:main"

[tool.coverage.run]
branch = true
source = ["trello_reports"]

[tool.coverage.report]
show_missing = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "nocov",
    "if TYPE_CHECKING:",
]
